# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Zinc Search engine API documents https://docs.zincsearch.com",
        "title": "Zinc Search engine API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Zinc Search",
            "url": "https://www.zincsearch.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:4080",
    "basePath": "/",
    "paths": {
...
        "/es/_msearch": {
            "post": {
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search V2 MultipleSearch for compatible ES",
                "operationId": "MSearch",
                "parameters": [
                    {
                        "description": "Query",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
        "/es/{index}/_search": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search V2 DSL for compatible ES",
                "operationId": "Search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Index",
                        "name": "index",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Query",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meta.ZincQueryForSDK"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meta.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/meta.HTTPResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "meta.AggregationAutoDateHistogram": {
            "type": "object",
            "properties": {
                "buckets": {
                    "type": "integer"
                },
                "field": {
                    "type": "string"
                },
                "format": {
                    "description": "format key_as_string",
                    "type": "string"
                },
                "keyed": {
                    "type": "boolean"
                },
                "minimum_interval": {
                    "description": "minute,hour,day,week,month,quarter,year",
                    "type": "string"
                },
                "time_zone": {
                    "description": "time_zone",
                    "type": "string"
                }
            }
        },
        "meta.AggregationDateHistogram": {
            "type": "object",
            "properties": {
                "calendar_interval": {
                    "description": "minute,hour,day,week,month,quarter,year",
                    "type": "string"
                },
                "extended_bounds": {
                    "$ref": "#/definitions/aggregation.HistogramBound"
                },
                "field": {
                    "type": "string"
                },
                "fixed_interval": {
                    "description": "ms,s,m,h,d",
                    "type": "string"
                },
                "format": {
                    "description": "format key_as_string",
                    "type": "string"
                },
                "hard_bounds": {
                    "$ref": "#/definitions/aggregation.HistogramBound"
                },
                "interval": {
                    "description": "ms,s,m,h,d",
                    "type": "string"
                },
                "keyed": {
                    "type": "boolean"
                },
                "min_doc_count": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "time_zone": {
                    "description": "time_zone",
                    "type": "string"
                }
            }
        },
        "meta.AggregationDateRange": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "format": {
                    "description": "format the `to` and `from` values to `_as_string`, and used to format `keyed response`",
                    "type": "string"
                },
                "keyed": {
                    "type": "boolean"
                },
                "ranges": {
                    "description": "refer",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meta.DateRange"
                    }
                },
                "time_zone": {
                    "description": "refer",
                    "type": "string"
                }
            }
        },
        "meta.AggregationHistogram": {
            "type": "object",
            "properties": {
                "extended_bounds": {
                    "$ref": "#/definitions/aggregation.HistogramBound"
                },
                "field": {
                    "type": "string"
                },
                "hard_bounds": {
                    "$ref": "#/definitions/aggregation.HistogramBound"
                },
                "interval": {
                    "type": "number"
                },
                "keyed": {
                    "type": "boolean"
                },
                "min_doc_count": {
                    "type": "integer"
                },
                "offset": {
                    "type": "number"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "meta.AggregationIPRange": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "keyed": {
                    "type": "boolean"
                },
                "ranges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meta.IPRange"
                    }
                }
            }
        },
        "meta.AggregationMetric": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "weight_field": {
                    "description": "Field name to be used for setting weight for primary field for weighted average aggregation",
                    "type": "string"
                }
            }
        },
        "meta.AggregationRange": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "keyed": {
                    "type": "boolean"
                },
                "ranges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meta.Range"
                    }
                }
            }
        },
        "meta.AggregationResponse": {
            "type": "object",
            "properties": {
                "buckets": {
                    "description": "slice or map"
                },
                "interval": {
                    "description": "support for auto_date_histogram_aggregation",
                    "type": "string"
                },
                "value": {}
            }
        },
        "meta.Aggregations": {
            "type": "object",
            "properties": {
                "aggs": {
                    "description": "nested aggregations",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.Aggregations"
                    }
                },
                "auto_date_histogram": {
                    "$ref": "#/definitions/meta.AggregationAutoDateHistogram"
                },
                "avg": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                },
                "cardinality": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                },
                "count": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                },
                "date_histogram": {
                    "$ref": "#/definitions/meta.AggregationDateHistogram"
                },
                "date_range": {
                    "$ref": "#/definitions/meta.AggregationDateRange"
                },
                "histogram": {
                    "$ref": "#/definitions/meta.AggregationHistogram"
                },
                "ip_range": {
                    "description": "TODO: not implemented",
                    "$ref": "#/definitions/meta.AggregationIPRange"
                },
                "max": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                },
                "min": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                },
                "range": {
                    "$ref": "#/definitions/meta.AggregationRange"
                },
                "sum": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                },
                "terms": {
                    "$ref": "#/definitions/meta.AggregationsTerms"
                },
                "weighted_avg": {
                    "$ref": "#/definitions/meta.AggregationMetric"
                }
            }
        },
        "meta.AggregationsTerms": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "order": {
                    "description": "{ \"_count\": \"asc\" }",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "meta.BoolQueryForSDK": {
            "type": "object",
            "properties": {
                "filter": {
                    "description": "query, [query1, query2]",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meta.QueryForSDK"
                    }
                },
                "minimum_should_match": {
                    "description": "only for should",
                    "type": "number"
                },
                "must": {
                    "description": "query, [query1, query2]",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meta.QueryForSDK"
                    }
                },
                "must_not": {
                    "description": "query, [query1, query2]",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meta.QueryForSDK"
                    }
                },
                "should": {
                    "description": "query, [query1, query2]",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meta.QueryForSDK"
                    }
                }
            }
        },
...
        "meta.ExistsQuery": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                }
            }
        },
        "meta.FuzzyQuery": {
            "type": "object",
            "properties": {
                "boost": {
                    "type": "number"
                },
                "fuzziness": {
                    "description": "auto, 1,2,3,n"
                },
                "prefix_length": {
                    "type": "number"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "meta.HTTPResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "meta.Highlight": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.Highlight"
                    }
                },
                "fragment_size": {
                    "type": "integer"
                },
                "number_of_fragments": {
                    "type": "integer"
                },
                "post_tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "pre_tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "meta.Hit": {
            "type": "object",
            "properties": {
                "@timestamp": {
                    "type": "string"
                },
                "_id": {
                    "type": "string"
                },
                "_index": {
                    "type": "string"
                },
                "_score": {
                    "type": "number"
                },
                "_source": {},
                "_type": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": true
                },
                "highlight": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "meta.Hits": {
            "type": "object",
            "properties": {
                "hits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meta.Hit"
                    }
                },
                "max_score": {
                    "type": "number"
                },
                "total": {
                    "$ref": "#/definitions/meta.Total"
                }
            }
        },
...
        "meta.IdsQuery": {
            "type": "object",
            "properties": {
                "values": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "meta.MatchAllQuery": {
            "type": "object"
        },
        "meta.MatchBoolPrefixQuery": {
            "type": "object",
            "properties": {
                "analyzer": {
                    "type": "string"
                },
                "boost": {
                    "type": "number"
                },
                "query": {
                    "type": "string"
                }
            }
        },
        "meta.MatchNoneQuery": {
            "type": "object"
        },
        "meta.MatchPhrasePrefixQuery": {
            "type": "object",
            "properties": {
                "analyzer": {
                    "type": "string"
                },
                "boost": {
                    "type": "number"
                },
                "query": {
                    "type": "string"
                }
            }
        },
        "meta.MatchPhraseQuery": {
            "type": "object",
            "properties": {
                "analyzer": {
                    "type": "string"
                },
                "boost": {
                    "type": "number"
                },
                "query": {
                    "type": "string"
                }
            }
        },
        "meta.MatchQuery": {
            "type": "object",
            "properties": {
                "analyzer": {
                    "type": "string"
                },
                "boost": {
                    "type": "number"
                },
                "fuzziness": {
                    "description": "auto, 1,2,3,n"
                },
                "operator": {
                    "description": "or(default), and",
                    "type": "string"
                },
                "prefix_length": {
                    "type": "number"
                },
                "query": {
                    "type": "string"
                }
            }
        },
        "meta.MultiMatchQuery": {
            "type": "object",
            "properties": {
                "analyzer": {
                    "type": "string"
                },
                "boost": {
                    "type": "number"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "minimum_should_match": {
                    "type": "number"
                },
                "operator": {
                    "description": "or(default), and",
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "type": {
                    "description": "best_fields(default), most_fields, cross_fields, phrase, phrase_prefix, bool_prefix",
                    "type": "string"
                }
            }
        },
...
        "meta.PrefixQuery": {
            "type": "object",
            "properties": {
                "boost": {
                    "type": "number"
                },
                "value": {
                    "description": "You can speed up prefix queries using the index_prefixes mapping parameter.",
                    "type": "string"
                }
            }
        },
...
        "meta.QueryForSDK": {
            "type": "object",
            "properties": {
                "bool": {
                    "description": ".",
                    "$ref": "#/definitions/meta.BoolQueryForSDK"
                },
                "exists": {
                    "description": ".",
                    "$ref": "#/definitions/meta.ExistsQuery"
                },
                "fuzzy": {
                    "description": "simple, PrefixQuery",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.FuzzyQuery"
                    }
                },
                "ids": {
                    "description": ".",
                    "$ref": "#/definitions/meta.IdsQuery"
                },
                "match": {
                    "description": "simple, MatchQuery",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.MatchQuery"
                    }
                },
                "match_all": {
                    "description": "just set or null",
                    "$ref": "#/definitions/meta.MatchAllQuery"
                },
                "match_bool_prefix": {
                    "description": "simple, MatchBoolPrefixQuery",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.MatchBoolPrefixQuery"
                    }
                },
                "match_none": {
                    "description": "just set or null",
                    "$ref": "#/definitions/meta.MatchNoneQuery"
                },
                "match_phrase": {
                    "description": "simple, MatchPhraseQuery",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.MatchPhraseQuery"
                    }
                },
                "match_phrase_prefix": {
                    "description": "simple, MatchPhrasePrefixQuery",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.MatchPhrasePrefixQuery"
                    }
                },
                "multi_match": {
                    "description": ".",
                    "$ref": "#/definitions/meta.MultiMatchQuery"
                },
                "prefix": {
                    "description": ".",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.PrefixQuery"
                    }
                },
                "query_string": {
                    "description": ".",
                    "$ref": "#/definitions/meta.QueryStringQuery"
                },
                "range": {
                    "description": "simple, FuzzyQuery",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.RangeQueryForSDK"
                    }
                },
                "regexp": {
                    "description": "simple, FuzzyQuery",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.RegexpQuery"
                    }
                },
                "simple_query_string": {
                    "description": ".",
                    "$ref": "#/definitions/meta.SimpleQueryStringQuery"
                },
                "term": {
                    "description": "simple, TermQuery",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.TermQueryForSDK"
                    }
                },
                "terms": {
                    "description": ".",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.TermsQuery"
                    }
                },
                "wildcard": {
                    "description": "simple, WildcardQuery",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.WildcardQuery"
                    }
                }
            }
        },
        "meta.QueryStringQuery": {
            "type": "object",
            "properties": {
                "analyzer": {
                    "type": "string"
                },
                "boost": {
                    "type": "number"
                },
                "default_field": {
                    "type": "string"
                },
                "default_operator": {
                    "description": "or(default), and",
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "query": {
                    "type": "string"
                }
            }
        },
...
        "meta.RangeQueryForSDK": {
            "type": "object",
            "properties": {
                "boost": {
                    "type": "number"
                },
                "format": {
                    "description": "Date format used to convert date values in the query.",
                    "type": "string"
                },
                "gt": {
                    "description": "string, float64",
                    "type": "string"
                },
                "gte": {
                    "description": "string, float64",
                    "type": "string"
                },
                "lt": {
                    "description": "string, float64",
                    "type": "string"
                },
                "lte": {
                    "description": "string, float64",
                    "type": "string"
                },
                "time_zone": {
                    "description": "used to convert date values in the query to UTC.",
                    "type": "string"
                }
            }
        },
        "meta.RegexpQuery": {
            "type": "object",
            "properties": {
                "boost": {
                    "type": "number"
                },
                "flags": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "meta.SearchResponse": {
            "type": "object",
            "properties": {
                "_shards": {
                    "$ref": "#/definitions/meta.Shards"
                },
                "aggregations": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.AggregationResponse"
                    }
                },
                "error": {
                    "type": "string"
                },
                "hits": {
                    "$ref": "#/definitions/meta.Hits"
                },
                "timed_out": {
                    "type": "boolean"
                },
                "took": {
                    "description": "Time it took to generate the response",
                    "type": "integer"
                }
            }
        },
        "meta.Shards": {
            "type": "object",
            "properties": {
                "failed": {
                    "type": "integer"
                },
                "skipped": {
                    "type": "integer"
                },
                "successful": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "meta.TermQueryForSDK": {
            "type": "object",
            "properties": {
                "boost": {
                    "type": "number"
                },
                "case_insensitive": {
                    "type": "boolean"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "meta.TermsQuery": {
            "type": "object",
            "additionalProperties": true
        },
        "meta.Total": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "Count of documents returned",
                    "type": "integer"
                }
            }
        },
...
        "meta.ZincQueryForSDK": {
            "type": "object",
            "properties": {
                "_source": {
                    "description": "true, false, [\"field1\", \"field2.*\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "aggs": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/meta.Aggregations"
                    }
                },
                "explain": {
                    "type": "boolean"
                },
                "fields": {
                    "description": "[\"field1\", \"field2.*\", {\"field\": \"fieldName\", \"format\": \"epoch_millis\"}]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "from": {
                    "type": "integer"
                },
                "highlight": {
                    "$ref": "#/definitions/meta.Highlight"
                },
                "query": {
                    "$ref": "#/definitions/meta.QueryForSDK"
                },
                "size": {
                    "type": "integer"
                },
                "sort": {
                    "description": "\"_sorce\", [\"+Year\",\"-Year\", {\"Year\": \"desc\"}, \"Date\": {\"order\": \"asc\"\", \"format\": \"yyyy-MM-dd\"}}\"}]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "timeout": {
                    "type": "integer"
                },
                "track_total_hits": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}